requirements.txt
flask==2.2.2
azure-storage-blob==12.15.0
pandas==1.5.3